% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alter_johnson.R
\name{alter_johnson}
\alias{alter_johnson}
\title{Alter-Johnson periodization}
\usage{
alter_johnson(y, t)
}
\arguments{
\item{y}{double. A vector of time series values.}

\item{t}{integer. An integer shift of time series values.}
}
\value{
A value of how likely t is a period (the closer to zero the better)
}
\description{
Generates an Alter-Johnson transformation function for provided time series.
The return function can be used for estimating a period in the data.
}
\examples{
x <- seq(-2 * pi, 2 * pi, length = 10001)
y <- sin(x)
alter_johnson(y, 5000)
# [1] 3.07982e-16
# which correlates to T = 6.270619 (almost 2 * pi)
}
