% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_periods.R
\name{find_periods}
\alias{find_periods}
\title{Find periods in time series using Alter-Johnson function.}
\usage{
find_periods(x, y, tol=1e-6)
}
\arguments{
\item{x}{double. A vector of time series keys.}

\item{y}{double. A vector of time series values.}

\item{tol}{double. A tolerance value for comparison between twe values.}
}
\value{
A vector of doubles – estimated periods.
}
\description{
Find periods in time series using Alter-Johnson function.
}
\examples{
x <- seq(-2 * pi, 2 * pi, length = 1001)
y <- sin(x)
find_periods(x, y)
# [1]  6.283185 12.566371
# 2*pi and 4*pi respectively
}
