% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrate_rect.R
\name{integrate_rect}
\alias{integrate_rect}
\title{Rectangular integration}
\usage{
integrate_rect(fn, left, right, n)
}
\arguments{
\item{fn}{A function that takes a number and returns another.}

\item{left}{Left boundary of integration.}

\item{right}{Right boundary of integration.}

\item{n}{Number of split intervals.}
}
\value{
An approximate value of the definite integral.
}
\description{
Integrate a function numerically using rectangular integration.
}
\examples{
fn <- function(x) x * x
integrate_rect(fn, 0, 4, 2000)
# 21.31734
}
