% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrate_simpson.R
\name{integrate_simpson}
\alias{integrate_simpson}
\title{Simpson's Rule integration}
\usage{
integrate_simpson(fn, left, right, n)
}
\arguments{
\item{fn}{A function that takes a number and returns another.}

\item{left}{Left boundary of integration.}

\item{right}{Right boundary of integration.}

\item{n}{Number of split intervals.}
}
\value{
An approximate value of the definite integral.
}
\description{
Integrate a function numerically using Simpson's Rule.
}
\examples{
integrate_rect(sin, 0, 3.14159, 2000)
# 2
}
