% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interp_newton.R
\name{interp_newton}
\alias{interp_newton}
\title{Newton polynomial}
\usage{
interp_newton(xs, ys)
}
\arguments{
\item{xs}{double. X values of interpolated function.}

\item{ys}{double. Y values of interpolated function.}
}
\value{
A function p(x) that calculates a value of the Newton polynomial
for given points at x.
}
\description{
Interpolate given points with a Newton polynomial.
}
\examples{
xs <- as.double(1:5)
p <- interp_newton(xs, sin(xs))
}
