% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linfit.R
\name{linfit}
\alias{linfit}
\title{Linear regression}
\usage{
linfit(xs, ys)
}
\arguments{
\item{xs}{double. Vector of dependent variable values.}

\item{ys}{double. Vector of independent variable values.}
}
\value{
List containing m and k.
}
\description{
Calculates coeffitients m and k for linear regression y = mx + k
via least squares method.
}
\examples{
fn <- function(x) 10 * x + 5
xs <- seq(1, 10, 0.5)
ys <- fn(xs)
linfit(xs, ys)
# $m
# [1] 10
# $k
# [1] 5
}
